poc-eco-classify/requirements.txt
sentence-transformers
transformers
torch
scikit-learn
pyyaml
